Move db-strategy into a more appropriate place, such as the Db config struct. Ensure the failing test doesn't fail.
