+++ .gitignore
@@ -2,3 +2,4 @@ target
 .idea
 .DS_Store
 /data
+.trustify
