+++ common/src/test_util.rs
@@ -1,10 +1,11 @@
-use crate::config::Database;
+use crate::config::{Database, DbStrategy};
 use crate::db;
 use std::sync::Arc;
 
 pub async fn bootstrap_system(name: &str) -> Result<db::Database, anyhow::Error> {
     db::Database::with_external_config(
         &Database {
+            db_strategy: DbStrategy::External,
             username: "postgres".to_string(),
             password: "eggs".to_string(),
             host: "localhost".to_string(),
